Move social user actions to common
