+++ src/components/artist/ArtistPopover.tsx
@@ -11,10 +11,10 @@ import { ID } from 'common/models/Identifiers'
 import { WidthSizes, SquareSizes } from 'common/models/ImageSizes'
 import { getUserId } from 'common/store/account/selectors'
 import { getUser } from 'common/store/cache/users/selectors'
+import * as socialActions from 'common/store/social/users/actions'
 import { MountPlacement } from 'components/types'
 import { setNotificationSubscription } from 'containers/profile-page/store/actions'
 import { useUserCoverPhoto, useUserProfilePicture } from 'hooks/useImageSize'
-import * as socialActions from 'store/social/users/actions'
 import { AppState } from 'store/types'
 import { profilePage } from 'utils/route'
