+++ src/components/artist/UserListModal.tsx
@@ -11,9 +11,9 @@ import loadingSpinner from 'assets/animations/loadingSpinner.json'
 import { FollowSource } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import User from 'common/models/User'
+import * as socialActions from 'common/store/social/users/actions'
 import ArtistChip from 'components/artist/ArtistChip'
 import FollowButton from 'components/general/FollowButton'
-import * as socialActions from 'store/social/users/actions'
 
 import styles from './UserListModal.module.css'
