+++ src/containers/artist-recommendations/ArtistRecommendations.tsx
@@ -15,6 +15,7 @@ import { FollowSource, Name } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import { ProfilePictureSizes, SquareSizes } from 'common/models/ImageSizes'
 import User from 'common/models/User'
+import * as socialActions from 'common/store/social/users/actions'
 import ArtistPopover from 'components/artist/ArtistPopover'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
 import FollowButton from 'components/general/FollowButton'
@@ -23,7 +24,6 @@ import { MountPlacement } from 'components/types'
 import UserBadges from 'containers/user-badges/UserBadges'
 import { useUserProfilePicture } from 'hooks/useImageSize'
 import { make, useRecord } from 'store/analytics/actions'
-import * as socialActions from 'store/social/users/actions'
 import { AppState } from 'store/types'
 import { useIsMobile } from 'utils/clientUtil'
 import { profilePage } from 'utils/route'
