+++ src/containers/collection-page/CollectionPageProvider.tsx
@@ -31,6 +31,7 @@ import {
   deletePlaylist
 } from 'common/store/cache/collections/actions'
 import * as socialTracksActions from 'common/store/social/tracks/actions'
+import * as socialUsersActions from 'common/store/social/users/actions'
 import { open } from 'common/store/ui/mobile-overflow-menu/actions'
 import {
   OverflowAction,
@@ -62,7 +63,6 @@ import { getPlaying, getBuffering } from 'store/player/selectors'
 import { makeGetCurrent } from 'store/queue/selectors'
 import { getLocationPathname } from 'store/routing/selectors'
 import * as socialCollectionsActions from 'store/social/collections/actions'
-import * as socialUsersActions from 'store/social/users/actions'
 import { AppState } from 'store/types'
 import {
   profilePage,
