+++ src/containers/feed-page/store/sagas.ts
@@ -3,11 +3,11 @@ import { call, put, take, fork, takeEvery } from 'redux-saga/effects'
 import { FollowSource } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
 import { fetchUsers } from 'common/store/cache/users/sagas'
+import * as socialActions from 'common/store/social/users/actions'
 import * as discoverActions from 'containers/feed-page/store/actions'
 import { feedActions } from 'containers/feed-page/store/lineups/feed/actions'
 import { fetchSuggestedFollowUserIds } from 'containers/sign-on/store/sagas'
 import { waitForBackendSetup } from 'store/backend/sagas'
-import * as socialActions from 'store/social/users/actions'
 
 import feedSagas from './lineups/feed/sagas'
