+++ src/containers/unfollow-confirmation-modal/UnfollowConfirmationModal.tsx
@@ -5,8 +5,8 @@ import { Dispatch } from 'redux'
 
 import { FollowSource } from 'common/models/Analytics'
 import { ID } from 'common/models/Identifiers'
+import * as socialActions from 'common/store/social/users/actions'
 import { setNotificationSubscription } from 'containers/profile-page/store/actions'
-import * as socialActions from 'store/social/users/actions'
 import { AppState } from 'store/types'
 
 import UnfollowConfirmationModal from './components/UnfollowConfirmationModal'
