+++ src/containers/user-list/UserList.tsx
@@ -9,9 +9,9 @@ import { ID } from 'common/models/Identifiers'
 import User from 'common/models/User'
 import { getUserId } from 'common/store/account/selectors'
 import { getUsers } from 'common/store/cache/users/selectors'
+import * as socialActions from 'common/store/social/users/actions'
 import { setNotificationSubscription } from 'containers/profile-page/store/actions'
 import * as unfollowConfirmationActions from 'containers/unfollow-confirmation-modal/store/actions'
-import * as socialActions from 'store/social/users/actions'
 import { AppState } from 'store/types'
 import { isMobile } from 'utils/clientUtil'
 import { profilePage } from 'utils/route'
