+++ src/store/social/users/errorSagas.ts
@@ -1,6 +1,6 @@
 import createErrorSagas from 'common/utils/errorSagas'
 
-import * as socialUserActions from './actions'
+import * as socialUserActions from '../../../common/store/social/users/actions'
 
 type UserErrors =
   | ReturnType<typeof socialUserActions.followUserFailed>
