+++ src/store/social/users/sagas.ts
@@ -7,6 +7,7 @@ import { getUserId } from 'common/store/account/selectors'
 import * as cacheActions from 'common/store/cache/actions'
 import { adjustUserField } from 'common/store/cache/users/sagas'
 import { getUsers, getUser } from 'common/store/cache/users/selectors'
+import * as socialActions from 'common/store/social/users/actions'
 import { makeKindId } from 'common/utils/uid'
 import * as signOnActions from 'containers/sign-on/store/actions'
 import AudiusBackend from 'services/AudiusBackend'
@@ -14,7 +15,6 @@ import { make } from 'store/analytics/actions'
 import { waitForBackendSetup } from 'store/backend/sagas'
 import * as confirmerActions from 'store/confirmer/actions'
 import { confirmTransaction } from 'store/confirmer/sagas'
-import * as socialActions from 'store/social/users/actions'
 import { profilePage } from 'utils/route'
 import { share } from 'utils/share'
