+++ src/store/social/users/store.test.js
@@ -5,8 +5,8 @@ import * as matchers from 'redux-saga-test-plan/matchers'
 import Kind from 'common/models/Kind'
 import * as cacheActions from 'common/store/cache/actions'
 import { adjustUserField } from 'common/store/cache/users/sagas'
+import * as actions from 'common/store/social/users/actions'
 import { waitForBackendSetup } from 'store/backend/sagas'
-import * as actions from 'store/social/users/actions'
 import * as sagas from 'store/social/users/sagas'
 import { noopReducer } from 'store/testHelper'
